refactor: 체크리스트 Mongo to RDB 마이그레이션
