+++ src/main/java/com/project/dorumdorum/domain/checklist/ui/LoadMyRoomRuleController.java
@@ -1,8 +1,8 @@
-package com.project.dorumdorum.domain.room.ui;
+package com.project.dorumdorum.domain.checklist.ui;
 
 import com.project.dorumdorum.domain.checklist.application.dto.response.MyRoomRuleResponse;
 import com.project.dorumdorum.domain.checklist.application.usecase.LoadMyRoomRuleUseCase;
-import com.project.dorumdorum.domain.room.ui.spec.LoadMyRoomRuleApiSpec;
+import com.project.dorumdorum.domain.checklist.ui.spec.LoadMyRoomRuleApiSpec;
 import com.project.dorumdorum.global.common.BaseResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PathVariable;
