+++ src/main/java/com/project/dorumdorum/domain/checklist/ui/spec/UpdateRoomRuleApiSpec.java
@@ -1,4 +1,4 @@
-package com.project.dorumdorum.domain.room.ui.spec;
+package com.project.dorumdorum.domain.checklist.ui.spec;
 
 import com.project.dorumdorum.domain.checklist.application.dto.request.UpdateRoomRuleRequest;
 import com.project.dorumdorum.global.common.BaseResponse;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@Tag(name = "Room")
+@Tag(name = "Checklist")
 public interface UpdateRoomRuleApiSpec {
 
     @Operation(
